use SilentAbortException; move test
