+++ universum/main.py
@@ -73,7 +73,9 @@ class Main(HasOutput):
         self.vcs.prepare_repository()
         project_configs: Configuration = self.launcher.process_project_configs()
         if not project_configs:
-            raise Exception("Project configs are empty, abort")
+            self.out.log("Project configs are empty, build skipped\n ")
+            self.out.report_build_status("Skipped - project configs are empty")
+            raise SilentAbortException(application_exit_code=1)
         afterall_configs: Configuration = self.code_report_collector.prepare_environment(project_configs)
         self.artifacts.set_and_clean_artifacts(project_configs)
